Moved auth to route group
