+++ app/api/auth/[...nextauth]/route.ts
@@ -96,7 +96,7 @@ export const authOptions: NextAuthOptions = {
     strategy: "jwt",
   },
   pages: {
-    signIn: "/auth/login",
+    signIn: "/login",
   },
 };
